+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes in **python-transip** are documented below.
 
 ## [Unreleased]
+
+## [0.5.0] (2021-02-10)
 ### Added
 - The option to replace all existing nameservers of a single domain at once from the `transip.v6.objects.Domain.nameservers` service.
 - The option to list all colocations from the `transip.TransIP.colocations` service ([#24](https://github.com/roaldnefs/python-transip/issues/24)).
@@ -19,5 +21,6 @@ All notable changes in **python-transip** are documented below.
 - The option to update the content of a single DNS record from the `transip.v6.objects.Domain.dns` service, as well as from the `transip.v6.objects.DnsEntry` object.
 - The option to replace all existing DNS records of a single domain at once from the `transip.v6.objects.Domain.dns` service.
 
-[Unreleased]: https://github.com/roaldnefs/python-transip/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/roaldnefs/python-transip/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/roaldnefs/python-transip/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/roaldnefs/python-transip/compare/v0.3.0...v0.4.0